This is the start of the clicker game project
This game starts as a simple clicker game

There is rules that will disappear after a set time

A username and password will eventually save progress

The times clicked will count as the first type of currency in the game.

This currency will be able to be spent on customization on a shop. 

This will be the first part on a bigger project where the game evolves into a board game.